Implemented plotting, removed redundant code
implemented 2 functions, one that plots the orbital trajectory during the mission, distance to glider and angle to glider, and another function that plots the trajectory of the capsule until atmospheric encounter and the trajectory of the orbiter untill the end of the atmospheric mission. atmospheric mission thing is bugged right now.